[feat/#103] - ChatSearch.tsx로 컴포넌트 분리 + 불필요한 주석 삭제
